+++ Haoji/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.wayney">
+    package="com.example.haoji">
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".dailyActivity.dailyActivity"
+            android:name="com.example.haoji.dailyActivity.dailyActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -23,36 +23,32 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".userActivity.login1Activity"
+            android:name="com.example.haoji.userActivity.login1Activity"
             android:label="@string/title_activity_login1"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name="userActivity.showinfoActivity"
+            android:name="com.example.haoji.userActivity.showinfoActivity"
             android:label="@string/title_activity_showinfo"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".userActivity.changePasswordActivity"
+            android:name="com.example.haoji.userActivity.changePasswordActivity"
             android:label="@string/title_activity_change_password"
             android:theme="@style/AppTheme.NoActionBar" />
 		 <activity
-            android:name=".userActivity.RegisterActivity"
+            android:name="com.example.haoji.userActivity.RegisterActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name=".dailyActivity.passChangeActivity"
+            android:name="com.example.haoji.dailyActivity.passChangeActivity"
             android:label="@string/title_activity_pass_change"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity
-            android:name=".dailyActivity.newPlan"
+            android:name="com.example.haoji.dailyActivity.newPlan"
             android:label="newPlan"
             android:theme="@style/AppTheme.NoActionBar"/>
         <activity
-            android:name=".userActivity.forgetPasswordActivity"
+            android:name="com.example.haoji.userActivity.forgetPasswordActivity"
             android:label="forgetPasswordActivit"
-            android:theme="@style/AppTheme.NoActionBar"/>
-        <activity
-            android:name=".userActivity.showinfoActivity"
-            android:label="showinfoActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
         <!-- 注册SDKActivity -->
         <activity
