+++ Haoji/app/src/main/java/com/example/haoji/Button/ScreenUtils.java
@@ -1,4 +1,4 @@
-package com.example.wayney.Button;
+package com.example.haoji.Button;
 
 /**
  * Created by WayneY on 2017/11/9.
