+++ Haoji/app/src/main/java/com/example/haoji/userActivity/changePasswordActivity.java
@@ -1,4 +1,4 @@
-package com.example.wayney.userActivity;
+package com.example.haoji.userActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,8 +8,8 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
 
-import com.example.wayney.R;
-import com.example.wayney.dailyActivity.passChangeActivity;
+import com.example.haoji.R;
+import com.example.haoji.dailyActivity.passChangeActivity;
 
 public class changePasswordActivity extends AppCompatActivity {
