+++ Haoji/app/src/main/res/layout/activity_login1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.wayney.login1Activity">
+    tools:context="com.example.haoji.login1Activity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
