+++ Haoji/app/src/main/res/layout/app_bar_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.wayney.dailyActivity.dailyActivity">
+    tools:context="com.example.haoji.dailyActivity.dailyActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
